The "Major performance boost!" update has been reverted.
After rebooting the server, the website stopped working and didn't wan't to load in any of the js or css files. I reverted the changes that I made yesterday and spend the whole day trying to find a fix. It's not perfect but at least it's faster that it used to be.